#100 [U-ways] Rename FileExt to PropertyClassLoader to better reflect utility purpose
