+++ propactive-plugin/src/main/kotlin/io/github/propactive/task/ValidateApplicationPropertiesTask.kt
@@ -6,7 +6,7 @@ import io.github.propactive.plugin.Configuration.Companion.DEFAULT_ENVIRONMENTS
 import io.github.propactive.plugin.Configuration.Companion.DEFAULT_IMPLEMENTATION_CLASS
 import io.github.propactive.plugin.Propactive.Companion.LOGGER
 import io.github.propactive.plugin.Propactive.Companion.PROPACTIVE_GROUP
-import io.github.propactive.task.support.load
+import io.github.propactive.task.support.PropertyClassLoader.load
 import org.gradle.api.tasks.TaskAction
 
 open class ValidateApplicationPropertiesTask : ApplicationPropertiesTask() {
